+++ forge/src/main/java/mod/bluestaggo/modernerbeta/forge/ModernerBetaForge.java
@@ -1,0 +1,58 @@
+package mod.bluestaggo.modernerbeta.forge;
+
+import dev.architectury.platform.forge.EventBuses;
+import mod.bluestaggo.modernerbeta.ModernerBeta;
+import net.minecraft.resource.*;
+import net.minecraft.text.Text;
+import net.minecraft.util.Formatting;
+import net.minecraftforge.event.AddPackFindersEvent;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.ModList;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+
+import java.nio.file.Path;
+
+@Mod(ModernerBeta.MOD_ID)
+@Mod.EventBusSubscriber(modid = ModernerBeta.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
+public class ModernerBetaForge {
+    public ModernerBetaForge() {
+        IEventBus bus = FMLJavaModLoadingContext.get().getModEventBus();
+        EventBuses.registerModEventBus(ModernerBeta.MOD_ID, bus);
+    }
+
+    @SubscribeEvent
+    public static void commonInit(FMLConstructModEvent event) {
+        ModernerBeta.init();
+    }
+
+    @SubscribeEvent
+    public static void addPackFinders(AddPackFindersEvent event) {
+        if (event.getPackType() == ResourceType.SERVER_DATA) {
+            Path resourcePath = ModList.get().getModFileById(ModernerBeta.MOD_ID).getFile().findResource("resourcepacks/reduced_height");
+            ResourcePackProfile pack = ResourcePackProfile.create(
+                    "moderner_beta/reduced_height",
+                    Text.of("Reduced Height"), //TODO: i18n perhaps?
+                    false,
+                    new DirectoryResourcePack.DirectoryBackedFactory(resourcePath, false),
+                    ResourceType.SERVER_DATA,
+                    ResourcePackProfile.InsertionPosition.TOP,
+                    new ResourcePackSource() {
+                        @Override
+                        public Text decorate(Text packDisplayName) {
+                            return Text.translatable("pack.nameAndSource", packDisplayName,
+                                    Text.translatable("pack.source.builtin")).formatted(Formatting.GRAY);
+                        }
+
+                        @Override
+                        public boolean canBeEnabledLater() {
+                            return false;
+                        }
+                    }
+            );
+            event.addRepositorySource(consumer -> consumer.accept(pack));
+        }
+    }
+}